requirements.txt
pandas
sqlalchemy
psycopg2-binary
jupyter
matplotlib
seaborn
